Installs ChakraUI; Updates fonts;
