+++ package.json
@@ -11,6 +11,10 @@
 		"prepare": "node -e \"if (process.env.NODE_ENV !== 'production'){process.exit(1)} \" || husky install"
 	},
 	"dependencies": {
+		"@chakra-ui/next-js": "^2.2.0",
+		"@chakra-ui/react": "^2.8.2",
+		"@emotion/react": "^11.11.3",
+		"@emotion/styled": "^11.11.0",
 		"@typescript-eslint/eslint-plugin": "^6.20.0",
 		"@typescript-eslint/parser": "^6.20.0",
 		"eslint-config-airbnb": "^19.0.4",
@@ -19,6 +23,7 @@
 		"eslint-plugin-jsx-a11y": "^6.8.0",
 		"eslint-plugin-react": "^7.33.2",
 		"eslint-plugin-react-hooks": "^4.6.0",
+		"framer-motion": "^11.0.3",
 		"next": "14.0.4",
 		"node": "^21.2.0",
 		"react": "18.2.0",
