--- shared/fonts/fonts.ts
@@ -1,5 +1,0 @@
-import localFont from 'next/font/local';
-
-export const magicRetro = localFont({ src: './Magic Retro.ttf' });
-
-export const alstoria = localFont({ src: './Alstoria.ttf' });
